Move logic of inheritable resource out of acl.js.
