+++ index.js
@@ -2,3 +2,4 @@ module.exports = require("./lib/acl.js");
 module.exports.redisBackend = require("./lib/redis-backend.js");
 module.exports.memoryBackend = require("./lib/memory-backend.js");
 module.exports.mongodbBackend = require("./lib/mongodb-backend.js");
+module.exports.inheritableResourceHandler = require("./lib/inheritable-resource-handler.js");
