+++ lib/acl.js
@@ -38,7 +38,7 @@ contract = require('./contract');
 
 contract.debug = true;
 
-var Acl = function (backend, logger, inheritanceStopper){
+var Acl = function (backend, logger, resourceHandler){
   contract(arguments)
     .params('object')
     .params('object','object')
@@ -46,7 +46,8 @@ var Acl = function (backend, logger, inheritanceStopper){
 
   this.logger = logger;
   this.backend = backend;
-  this.inheritanceStopper = inheritanceStopper || '-';
+  
+  this.resourceHandler = resourceHandler || new Acl.inheritableResourceHandler();
 
   // Promisify async methods
   backend.endAsync = bluebird.promisify(backend.end);
@@ -418,34 +419,14 @@ Acl.prototype.isAllowed = function(userId, resource, permissions, cb){
 
   var _this = this;
   
-  var inheritanceStopperTester = new RegExp(this.inheritanceStopper);
-  
   if (Array.isArray(resource)) {
-    return this.allowedPermissions(userId, resource).then(function(resourcePermissions) {
-      var mergedPermissions = [];
-      _.forEach(resourcePermissions, function(resPermissions) {
-        // remove non-inheritable permissions obtained from previous level
-        mergedPermissions = _.filter(mergedPermissions, function(p) { return !inheritanceStopperTester.test(p); });
-        // add new resource permissions
-        mergedPermissions = _.uniq(_.compact(_.union(mergedPermissions, resPermissions)));
-      });
-      // remove inheritanceStopper from permissions
-      mergedPermissions = _.map(mergedPermissions, function(p) {
-        if (inheritanceStopperTester.test(p)) {
-          return p.substr(0, p.length - _this.inheritanceStopper.length);
-        } else {
-          return p;
-        }
-      });
-
-      if (mergedPermissions.indexOf('*') !== -1){
-        return true;
-      }
-
-      return makeArray(permissions).filter(function(p){
-        return mergedPermissions.indexOf(p) === -1;
-      }).length === 0;
-    }).nodeify(cb);
+    if (!_this.resourceHandler) {
+      cb(new Error('resourceHandler is required to support inheritable resource.'));
+    } else {
+      return this.allowedPermissions(userId, resource).then(function(resourcePermissions) {
+        return _this.resourceHandler.isAllowed(permissions, resourcePermissions);
+      }).nodeify(cb);
+    }
   } else {
     return this.backend.getAsync('users', userId).then(function(roles){
       if(roles.length){
