+++ lib/inheritable-resource-handler.js
@@ -1,0 +1,60 @@
+/**
+	Simple Resource Builder.
+	
+  Implementation of the resource ID builder using part of URL
+*/
+"use strict";
+
+var    _ = require('lodash'),
+contract = require('./contract');
+
+function InheritableResourceHandler(inheritanceStopper){  
+  this.inheritanceStopper = inheritanceStopper || '-';
+}
+
+InheritableResourceHandler.prototype = {
+  
+  isAllowed : function(permissions, resourcePermissions) {
+    contract(arguments).params('string|array', 'object').end();
+    
+    var _this = this;
+    var inheritanceStopperTester = new RegExp(this.inheritanceStopper);
+    var mergedPermissions = [];
+    _.forEach(resourcePermissions, function(resPermissions) {
+     // remove non-inheritable permissions obtained from previous level
+      mergedPermissions = _.filter(mergedPermissions, function(p) { return !inheritanceStopperTester.test(p); });
+      // add new resource permissions
+      mergedPermissions = _.uniq(_.compact(_.union(mergedPermissions, resPermissions)));
+    });
+
+    // remove inheritanceStopper from permissions
+    mergedPermissions = _.map(mergedPermissions, function(p) {
+      if (inheritanceStopperTester.test(p)) {
+        return p.substr(0, p.length - _this.inheritanceStopper.length);
+      } else {
+        return p;
+      }
+    });
+
+    if (mergedPermissions.indexOf('*') !== -1){
+      return true;
+    }
+
+    return makeArray(permissions).filter(function(p){
+      return mergedPermissions.indexOf(p) === -1;
+    }).length === 0;
+  }
+}
+
+//-----------------------------------------------------------------------------
+//
+// Helpers
+//
+//-----------------------------------------------------------------------------
+
+function makeArray(arr){
+  return Array.isArray(arr) ? arr : [arr];
+}
+
+exports = module.exports = InheritableResourceHandler;
+
